Remove `lib/` shifting.
Move `packages/*/lib/index.js` to `packages/*/index.js`.
