+++ packages/container/index.js
@@ -5,5 +5,5 @@ The public API, specified on the application namespace should be considered the 
   @private
 */
 
-export { default as Registry, privatize } from './registry';
-export { default as Container, FACTORY_FOR } from './container';
+export { default as Registry, privatize } from './lib/registry';
+export { default as Container, FACTORY_FOR } from './lib/container';
