+++ packages/ember-extension-support/index.js
@@ -1,0 +1,2 @@
+export { default as DataAdapter } from './lib/data_adapter';
+export { default as ContainerDebugAdapter } from './lib/container_debug_adapter';
