--- packages/ember-extension-support/lib/index.js
@@ -1,2 +1,0 @@
-export { default as DataAdapter } from './data_adapter';
-export { default as ContainerDebugAdapter } from './container_debug_adapter';
