+++ packages/ember-metal/index.js
@@ -1,0 +1,102 @@
+export {
+  default as computed,
+  getCacheFor,
+  getCachedValueFor,
+  peekCacheFor,
+  ComputedProperty,
+  _globalsComputed,
+} from './lib/computed';
+export { default as alias } from './lib/alias';
+export { default as merge } from './lib/merge';
+export { deprecateProperty } from './lib/deprecate_property';
+export {
+  instrument,
+  flaggedInstrument,
+  _instrumentStart,
+  reset as instrumentationReset,
+  subscribe as instrumentationSubscribe,
+  unsubscribe as instrumentationUnsubscribe,
+} from './lib/instrumentation';
+export { getOnerror, setOnerror, setDispatchOverride, getDispatchOverride } from './lib/error_handler';
+export { descriptorFor, meta, peekMeta, deleteMeta } from './lib/meta';
+export { default as Cache } from './lib/cache';
+export { PROXY_CONTENT, _getPath, get, getWithDefault } from './lib/property_get';
+export { set, trySet } from './lib/property_set';
+export {
+  objectAt,
+  replace,
+  replaceInNativeArray,
+  addArrayObserver,
+  removeArrayObserver,
+  arrayContentWillChange,
+  arrayContentDidChange,
+} from './lib/array';
+export { eachProxyFor, eachProxyArrayWillChange, eachProxyArrayDidChange } from './lib/each_proxy';
+export { addListener, hasListeners, on, removeListener, sendEvent } from './lib/events';
+
+export { default as isNone } from './lib/is_none';
+export { default as isEmpty } from './lib/is_empty';
+export { default as isBlank } from './lib/is_blank';
+export { default as isPresent } from './lib/is_present';
+export {
+  getCurrentRunLoop,
+  backburner,
+  run,
+  join,
+  bind,
+  begin,
+  end,
+  schedule,
+  hasScheduledTimers,
+  cancelTimers,
+  later,
+  once,
+  scheduleOnce,
+  next,
+  cancel,
+  debounce,
+  throttle,
+  _globalsRun,
+} from './lib/run_loop';
+export {
+  beginPropertyChanges,
+  changeProperties,
+  endPropertyChanges,
+  notifyPropertyChange,
+  overrideChains,
+  propertyDidChange,
+  propertyWillChange,
+  PROPERTY_DID_CHANGE,
+} from './lib/property_events';
+export { defineProperty, Descriptor } from './lib/properties';
+export { watchKey, unwatchKey } from './lib/watch_key';
+export { ChainNode, finishChains, removeChainWatcher } from './lib/chains';
+export { watchPath, unwatchPath } from './lib/watch_path';
+export { isWatching, unwatch, watch, watcherCount } from './lib/watching';
+export { default as libraries, Libraries } from './lib/libraries';
+export { Map, MapWithDefault, OrderedSet } from './lib/map';
+export { default as getProperties } from './lib/get_properties';
+export { default as setProperties } from './lib/set_properties';
+export { default as expandProperties } from './lib/expand_properties';
+
+export { addObserver, removeObserver } from './lib/observer';
+export { Mixin, aliasMethod, mixin, observer } from './lib/mixin';
+export { default as InjectedProperty } from './lib/injected_property';
+export { setHasViews, tagForProperty, tagFor, markObjectAsDirty } from './lib/tags';
+export { default as runInTransaction, didRender, assertNotRendered } from './lib/transaction';
+export { default as descriptor } from './lib/descriptor';
+export { tracked } from './lib/tracked';
+
+export {
+  NAMESPACES,
+  NAMESPACES_BY_ID,
+  addNamespace,
+  classToString,
+  findNamespace,
+  findNamespaces,
+  processNamespace,
+  processAllNamespaces,
+  removeNamespace,
+  isSearchDisabled as isNamespaceSearchDisabled,
+  setSearchDisabled as setNamespaceSearchDisabled,
+} from './lib/namespace_search';
