--- packages/ember-metal/lib/index.js
@@ -1,102 +1,0 @@
-export {
-  default as computed,
-  getCacheFor,
-  getCachedValueFor,
-  peekCacheFor,
-  ComputedProperty,
-  _globalsComputed,
-} from './computed';
-export { default as alias } from './alias';
-export { default as merge } from './merge';
-export { deprecateProperty } from './deprecate_property';
-export {
-  instrument,
-  flaggedInstrument,
-  _instrumentStart,
-  reset as instrumentationReset,
-  subscribe as instrumentationSubscribe,
-  unsubscribe as instrumentationUnsubscribe,
-} from './instrumentation';
-export { getOnerror, setOnerror, setDispatchOverride, getDispatchOverride } from './error_handler';
-export { descriptorFor, meta, peekMeta, deleteMeta } from './meta';
-export { default as Cache } from './cache';
-export { PROXY_CONTENT, _getPath, get, getWithDefault } from './property_get';
-export { set, trySet } from './property_set';
-export {
-  objectAt,
-  replace,
-  replaceInNativeArray,
-  addArrayObserver,
-  removeArrayObserver,
-  arrayContentWillChange,
-  arrayContentDidChange,
-} from './array';
-export { eachProxyFor, eachProxyArrayWillChange, eachProxyArrayDidChange } from './each_proxy';
-export { addListener, hasListeners, on, removeListener, sendEvent } from './events';
-
-export { default as isNone } from './is_none';
-export { default as isEmpty } from './is_empty';
-export { default as isBlank } from './is_blank';
-export { default as isPresent } from './is_present';
-export {
-  getCurrentRunLoop,
-  backburner,
-  run,
-  join,
-  bind,
-  begin,
-  end,
-  schedule,
-  hasScheduledTimers,
-  cancelTimers,
-  later,
-  once,
-  scheduleOnce,
-  next,
-  cancel,
-  debounce,
-  throttle,
-  _globalsRun,
-} from './run_loop';
-export {
-  beginPropertyChanges,
-  changeProperties,
-  endPropertyChanges,
-  notifyPropertyChange,
-  overrideChains,
-  propertyDidChange,
-  propertyWillChange,
-  PROPERTY_DID_CHANGE,
-} from './property_events';
-export { defineProperty, Descriptor } from './properties';
-export { watchKey, unwatchKey } from './watch_key';
-export { ChainNode, finishChains, removeChainWatcher } from './chains';
-export { watchPath, unwatchPath } from './watch_path';
-export { isWatching, unwatch, watch, watcherCount } from './watching';
-export { default as libraries, Libraries } from './libraries';
-export { Map, MapWithDefault, OrderedSet } from './map';
-export { default as getProperties } from './get_properties';
-export { default as setProperties } from './set_properties';
-export { default as expandProperties } from './expand_properties';
-
-export { addObserver, removeObserver } from './observer';
-export { Mixin, aliasMethod, mixin, observer } from './mixin';
-export { default as InjectedProperty } from './injected_property';
-export { setHasViews, tagForProperty, tagFor, markObjectAsDirty } from './tags';
-export { default as runInTransaction, didRender, assertNotRendered } from './transaction';
-export { default as descriptor } from './descriptor';
-export { tracked } from './tracked';
-
-export {
-  NAMESPACES,
-  NAMESPACES_BY_ID,
-  addNamespace,
-  classToString,
-  findNamespace,
-  findNamespaces,
-  processNamespace,
-  processAllNamespaces,
-  removeNamespace,
-  isSearchDisabled as isNamespaceSearchDisabled,
-  setSearchDisabled as setNamespaceSearchDisabled,
-} from './namespace_search';
