--- packages/ember-views/lib/index.js
@@ -1,28 +1,0 @@
-export { default as jQuery, jQueryDisabled } from './system/jquery';
-export {
-  addChildView,
-  isSimpleClick,
-  getViewBounds,
-  getViewClientRects,
-  getViewBoundingClientRect,
-  getRootViews,
-  getChildViews,
-  getViewId,
-  getViewElement,
-  setViewElement,
-  constructStyleDeprecationMessage,
-} from './system/utils';
-export { default as EventDispatcher } from './system/event_dispatcher';
-export { default as ComponentLookup } from './component_lookup';
-export { default as TextSupport } from './mixins/text_support';
-export { default as CoreView } from './views/core_view';
-export { default as ClassNamesSupport } from './mixins/class_names_support';
-export { default as ChildViewsSupport } from './mixins/child_views_support';
-export { default as ViewStateSupport } from './mixins/view_state_support';
-export { default as ViewMixin } from './mixins/view_support';
-export { default as ActionSupport } from './mixins/action_support';
-export { MUTABLE_CELL } from './compat/attrs';
-export { default as lookupPartial, hasPartial } from './system/lookup_partial';
-export { default as lookupComponent } from './utils/lookup-component';
-export { default as ActionManager } from './system/action_manager';
-export { default as fallbackViewRegistry } from './compat/fallback-view-registry';
