--- packages/internal-test-helpers/lib/index.js
@@ -1,30 +1,0 @@
-export { default as factory } from './factory';
-export { default as buildOwner } from './build-owner';
-export { default as confirmExport } from './confirm-export';
-export { default as equalInnerHTML } from './equal-inner-html';
-export { default as equalTokens } from './equal-tokens';
-export { default as moduleFor } from './module-for';
-export { default as strip } from './strip';
-export { default as applyMixins } from './apply-mixins';
-export { default as getTextOf } from './get-text-of';
-export { equalsElement, classes, styles, regex } from './matchers';
-export { runAppend, runDestroy } from './run';
-
-export { default as AbstractTestCase } from './test-cases/abstract';
-export { default as AbstractApplicationTestCase } from './test-cases/abstract-application';
-export { default as ApplicationTestCase } from './test-cases/application';
-export { default as QueryParamTestCase } from './test-cases/query-param';
-export { default as AbstractRenderingTestCase } from './test-cases/abstract-rendering';
-export { default as RenderingTestCase } from './test-cases/rendering';
-export { default as RouterTestCase } from './test-cases/router';
-export { default as AutobootApplicationTestCase } from './test-cases/autoboot-application';
-export {
-  default as DefaultResolverApplicationTestCase,
-} from './test-cases/default-resolver-application';
-
-export {
-  default as TestResolver,
-  ModuleBasedResolver as ModuleBasedTestResolver,
-} from './test-resolver';
-
-export { isIE11 } from './browser-detect';
