+++ tests/index.html
@@ -140,8 +140,8 @@
         window.EmberDev = window.EmberDev || {};
         EmberDev.runningProdBuild = !!QUnit.urlParams.prod;
 
-        var EmberDevTestHelperAssert = window.Ember.__loader.require('internal-test-helpers/ember-dev/index')['default'];
-        var setupQUnit = window.Ember.__loader.require('internal-test-helpers/ember-dev/setup-qunit')['default'];
+        var EmberDevTestHelperAssert = window.Ember.__loader.require('internal-test-helpers/lib/ember-dev/index')['default'];
+        var setupQUnit = window.Ember.__loader.require('internal-test-helpers/lib/ember-dev/setup-qunit')['default'];
         var debugModule = window.Ember.__loader.require('ember-debug');
         var testHelpers = new EmberDevTestHelperAssert({
           Ember: window.Ember,
